requirements.txt
beautifulsoup4==4.13.5
numpy==2.3.2
pandas==2.3.2
pytest==8.4.1
python-telegram-bot==22.3
Requests==2.32.5
yfinance==0.2.65
